Some more if clauses

+++ cleanup.py
@@ -11,7 +11,7 @@ def cleanup(mountpoint, tc_mountpoint, tempdir):
 # Unmounting Truecrypt
 #
     try:
-        if os.path.ismount(tc_mountpoint):
+        if os.path.exists(tc_mountpoint) and os.path.ismount(tc_mountpoint):
             mainLogger.info("Unmounting Truecrypt Container")
 	    mainLogger.debug('UNMOUNT COMMAND: ' + parser.get('truecrypting', 'unmount'))
 	    subprocess.check_call(shlex.split(parser.get('truecrypting', 'unmount')))
@@ -24,7 +24,7 @@ def cleanup(mountpoint, tc_mountpoint, tempdir):
 #
 
     try:
-        if args.device and os.path.ismount(mountpoint):
+        if args.device and os.path.exists(mountpoint) and os.path.ismount(mountpoint) == True:
             mainLogger.info("Unmounting USB-Stick")
             if (sys.platform == "darwin"):
                 subprocess.check_call(['diskutil', 'eject', mountpoint])
@@ -41,10 +41,14 @@ def cleanup(mountpoint, tc_mountpoint, tempdir):
     mainLogger.info("Cleaning up Temporary Directories")
 
     try:
-        if args.device and os.path.exists(mountpoint):
+        if args.device and os.path.exists(mountpoint) and os.path.ismount(mountpoint) == False:
             shutil.rmtree(mountpoint)
-        if os.path.exists(tc_mountpoint):
+        if os.path.exists(tc_mountpoint) and os.path.ismount(tc_mountpoint) == False:
             shutil.rmtree(tc_mountpoint)
+        if (sys.platform == "darwin") and if os.path.exists(tempdir+"/dmg"):
+            subprocess.check_call(['diskutil', 'eject', os.path.join(tempdir+"/dmg/")])
+        else:
+            subprocess.check_call(['umount', os.path.join(tempdir+"/dmg/")])
         if os.path.exists(tempdir):
             shutil.rmtree(tempdir)
     except OSError:
